+++ packages/sst/src/bootstrap.ts
@@ -261,7 +261,7 @@ export async function bootstrapSST(cdkBucket: string) {
     ),
     handler: "index.handler",
     runtime: region?.startsWith("us-gov-")
-      ? Runtime.NODEJS_16_X
+      ? Runtime.NODEJS_18_X
       : Runtime.NODEJS_18_X,
     environment: {
       BUCKET_NAME: bucket.bucketName,