+++ packages/sst/src/constructs/NextjsSite.ts
@@ -101,6 +101,7 @@ export interface NextjsSiteProps extends Omit<SsrSiteProps, "nodejs"> {
      * experimental: {
      *   disableIncrementalCache: true,
      * }
+     * ```
      */
     disableIncrementalCache?: boolean;
     /**
@@ -858,8 +859,7 @@ export class NextjsSite extends SsrSite {
         });
 
         useFunctions().sourcemaps.add(stack.stackName, {
-          srcBucket: asset.bucket,
-          srcKey: asset.s3ObjectKey,
+          asset,
           tarKey: path.join(server.functionArn, sourcemapKey),
         });
       });