+++ packages/sst/src/constructs/SsrFunction.ts
@@ -60,7 +60,7 @@ export interface SsrFunctionProps
   > {
   bundle?: string;
   handler: string;
-  runtime?: "nodejs14.x" | "nodejs16.x" | "nodejs18.x";
+  runtime?: "nodejs16.x" | "nodejs18.x";
   architecture?: "arm_64" | "x86_64" | Architecture;
   timeout?: number | Duration;
   memorySize?: number | Size;
@@ -185,12 +185,8 @@ export class SsrFunction extends Construct implements SSTConstruct {
         assetKey
       ),
       runtime:
-        runtime === "nodejs14.x"
-          ? Runtime.NODEJS_14_X
-          : runtime === "nodejs16.x"
-          ? Runtime.NODEJS_16_X
-          : Runtime.NODEJS_18_X,
-      architecture:
+        runtime === "nodejs16.x" ? Runtime.NODEJS_16_X : Runtime.NODEJS_18_X,
+        architecture:
         architecture instanceof Architecture
           ? architecture
           : architecture === "x86_64"
@@ -320,8 +316,7 @@ export class SsrFunction extends Construct implements SSTConstruct {
       });
       await fs.rm(result.sourcemap);
       useFunctions().sourcemaps.add(stack.stackName, {
-        srcBucket: asset.bucket,
-        srcKey: asset.s3ObjectKey,
+        asset,
         tarKey: this.functionArn,
       });
     }